Been awhile since I've commited
Ratio clustering is implemented and fixed in ipy notebook. Don't trust the LLR clustering- I'm doing what turns out to be a shady transformation there (not a good idea to just shift everything to zero).
Also got a start on some sensitivity specificity plots. Still a lot of good feedback to incorporate there.

Got a better record of my datasets for a rerun of all data. Created a summary table for the paper. Went excel->csv->ipython(pandas)->tex. It's a silly chain but gosh it worked. Much faster than manual table building!